Business/GROWTH/executions/Marketing/linkedin_bot_requirements.txt
APScheduler>=3.10.0
fastapi>=0.115.0
httpx>=0.27.0
notion-client>=3.0.0
playwright==1.58.0
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn[standard]>=0.30.0
